Move translation resources and logic into the same bundle.
