+++ chemclipse/plugins/org.eclipse.chemclipse.processing/src/org/eclipse/chemclipse/processing/methods/AbstractProcessEntryContainer.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Lablicate GmbH.
+ * Copyright (c) 2021, 2024 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
